backend/requirements.txt
Flask
Flask-Cors
pymongo
python-dotenv
python-docx
reportlab
PyMuPDF
gunicorn